Almost done backend

+++ src/main/java/com/wpproject/theater/repositories/ActorRepository.java
@@ -3,5 +3,5 @@ package com.wpproject.theater.repositories;
 import com.wpproject.theater.models.Actor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ActorRepository extends JpaRepository<Actor, Integer> {
+public interface ActorRepository extends JpaRepository<Actor, Long> {
 }
